app/requirements.txt
kafka-python
streamlit
streamlit-autorefresh
minio
pandas